Refactor node architecture for consistent naming conventions
- Rename TextNodeViewer → TextNode (node wrapper)
- Rename DateNodeViewer → DateNode (node wrapper)
- Rename DatePageViewer → DateNodeViewer (viewer wrapper)
- Update all imports and references across codebase
- Remove legacy TextNode wrapper component
- Update CSS class names and component comments
- Establish clear architectural pattern:
  * *Node = wraps BaseNode
  * *NodeViewer = wraps BaseNodeViewer

All type checks pass with 0 errors/warnings.
