+++ packages/desktop-app/src/lib/components/viewers/date-node-viewer.svelte
@@ -1,96 +1,216 @@
 <!--
-  DateNodeViewer - Node-level viewer for date content
-  
-  This is a simple wrapper around BaseNode that provides date-specific formatting
-  and validation for date-type nodes in the node tree.
+  DateNodeViewer - Specialized viewer that wraps BaseNodeViewer with date navigation
+
+  Features:
+  - Date navigation header with professional styling
+  - Previous/next day functionality
+  - Keyboard support for arrow key navigation
+  - Seamless integration with existing BaseNodeViewer
+  - Clean date formatting (e.g., "September 7, 2025")
 -->
 
 <script lang="ts">
-  import { createEventDispatcher } from 'svelte';
-  import BaseNode from '$lib/design/components/base-node.svelte';
+  import BaseNodeViewer from '$lib/design/components/base-node-viewer.svelte';
   import Icon from '$lib/design/icons/icon.svelte';
-  import type { NodeViewerProps } from '$lib/types/nodeViewers.js';
+  import { updateTabTitle, getDateTabTitle } from '$lib/stores/navigation.js';
+  import NodeServiceContext from '$lib/contexts/node-service-context.svelte';
 
-  // Props following the NodeViewer interface
-  let {
-    nodeId,
-    content = '',
-    autoFocus = false,
-    nodeType = 'date',
-    inheritHeaderLevel = 0,
-    children = []
-  }: NodeViewerProps = $props();
+  // Props using Svelte 5 runes mode
+  let { tabId = 'today' }: { tabId?: string } = $props();
 
-  const dispatch = createEventDispatcher();
+  // Current date state - defaults to today
+  let currentDate = $state(new Date());
 
-  // Parse date from content - expects YYYY-MM-DD format or similar
-  function parseDate(content: string): Date | null {
-    if (!content.trim()) return new Date(); // Default to today if empty
-
-    const dateMatch = content.match(/^\d{4}-\d{2}-\d{2}/);
-    if (dateMatch) {
-      return new Date(dateMatch[0]);
-    }
-
-    // Try to parse as a general date
-    const parsed = Date.parse(content.trim());
-    return isNaN(parsed) ? null : new Date(parsed);
-  }
-
-  // Format date for display
-  const parsedDate = $derived(parseDate(content));
+  // Format date for display (e.g., "September 7, 2025") using Svelte 5 $derived
   const formattedDate = $derived(
-    parsedDate
-      ? parsedDate.toLocaleDateString('en-US', {
-          year: 'numeric',
-          month: 'long',
-          day: 'numeric'
-        })
-      : content
+    currentDate.toLocaleDateString('en-US', {
+      year: 'numeric',
+      month: 'long',
+      day: 'numeric'
+    })
   );
 
-  // Enhanced content with date formatting
-  const displayContent = $derived(parsedDate ? `📅 ${formattedDate}` : content);
+  // Update tab title when date changes using Svelte 5 $effect
+  $effect(() => {
+    if (tabId) {
+      const newTitle = getDateTabTitle(currentDate);
+      updateTabTitle(tabId, newTitle);
+    }
+  });
+
+  /**
+   * Navigate to previous or next day
+   */
+  function navigateDate(direction: 'prev' | 'next') {
+    const newDate = new Date(currentDate);
+    if (direction === 'prev') {
+      newDate.setDate(newDate.getDate() - 1);
+    } else {
+      newDate.setDate(newDate.getDate() + 1);
+    }
+    currentDate = newDate;
+  }
+
+  /**
+   * Handle keyboard navigation for date switching
+   */
+  function handleKeydown(event: KeyboardEvent) {
+    // Only handle if focus is not on contenteditable elements
+    const activeElement = document.activeElement as HTMLElement;
+    const isTextEditor = activeElement && activeElement.contentEditable === 'true';
+
+    if (isTextEditor) return;
+
+    if (event.key === 'ArrowLeft') {
+      event.preventDefault();
+      navigateDate('prev');
+    } else if (event.key === 'ArrowRight') {
+      event.preventDefault();
+      navigateDate('next');
+    }
+  }
 </script>
 
+<!-- Keyboard event listener -->
+<svelte:window on:keydown={handleKeydown} />
+
 <div class="date-node-viewer">
-  <div class="date-icon">
-    <Icon name="calendar" size={16} color="hsl(var(--muted-foreground))" />
+  <!-- Date Navigation Header - matches original +page.svelte design -->
+  <div class="date-header">
+    <div class="date-nav-container">
+      <div class="date-display">
+        <Icon
+          name="calendar"
+          size={20}
+          color="hsl(var(--muted-foreground))"
+          className="calendar-icon"
+        />
+        <h1
+          style="font-size: 2rem; font-weight: 500; color: hsl(var(--muted-foreground)); margin: 0;"
+        >
+          {formattedDate}
+        </h1>
+      </div>
+
+      <div class="date-nav-buttons">
+        <button class="date-nav-btn" onclick={() => navigateDate('prev')} aria-label="Previous day">
+          <Icon name="chevronRight" size={16} color="currentColor" className="rotate-left" />
+        </button>
+        <button class="date-nav-btn" onclick={() => navigateDate('next')} aria-label="Next day">
+          <Icon name="chevronRight" size={16} color="currentColor" />
+        </button>
+      </div>
+    </div>
   </div>
-  <BaseNode
-    {nodeId}
-    content={displayContent}
-    {autoFocus}
-    {nodeType}
-    headerLevel={inheritHeaderLevel}
-    {children}
-    on:createNewNode
-    on:contentChanged={(e) => {
-      // Strip formatting when content changes
-      const rawContent = e.detail.content.replace(/^📅\s*/, '');
-      dispatch('contentChanged', { content: rawContent });
-    }}
-    on:indentNode
-    on:outdentNode
-    on:navigateArrow
-    on:combineWithPrevious
-    on:deleteNode
-    on:focus
-    on:blur
-  />
+
+  <!-- Node Content Area -->
+  <div class="node-content-area">
+    <NodeServiceContext>
+      <BaseNodeViewer />
+    </NodeServiceContext>
+  </div>
 </div>
 
 <style>
   .date-node-viewer {
+    /* Full height container for the entire date viewer */
     display: flex;
-    align-items: flex-start;
+    flex-direction: column;
+    height: 100vh;
+    max-height: 100vh;
+    width: 100%;
+    overflow: hidden;
+  }
+
+  /* Date header styling - matches original +page.svelte design exactly */
+  .date-header {
+    border-bottom: 1px solid hsl(var(--border));
+    background: hsl(var(--background));
+    flex-shrink: 0;
+  }
+
+  .date-nav-container {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 1rem;
+  }
+
+  .date-display {
+    display: flex;
+    align-items: center;
     gap: 0.5rem;
   }
 
-  .date-icon {
+  .date-nav-buttons {
     display: flex;
     align-items: center;
-    margin-top: 0.125rem; /* Align with text baseline */
-    flex-shrink: 0;
+    gap: 0.5rem;
+  }
+
+  .date-nav-btn {
+    width: 32px;
+    height: 32px;
+    border: 1px solid hsl(var(--border));
+    border-radius: 0.375rem;
+    background: hsl(var(--background));
+    color: hsl(var(--muted-foreground));
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    transition: all 0.15s ease;
+  }
+
+  .date-nav-btn:hover {
+    background: hsl(var(--muted));
+    color: hsl(var(--foreground));
+  }
+
+  .node-content-area {
+    /* Container for the BaseNodeViewer - takes remaining space */
+    flex: 1;
+    min-height: 0; /* Allow flex item to shrink below its content size */
+    overflow-y: auto;
+    overflow-x: hidden;
+    display: flex;
+    flex-direction: column;
+    /* Add proper padding for node tree spacing - matches original editor-area */
+    padding: 1.5rem;
+    /* Add extra bottom padding for better visibility of last node */
+    padding-bottom: 3rem;
+    /* Ensure smooth scrolling */
+    scroll-behavior: smooth;
+    /* Add subtle scrollbar styling */
+    scrollbar-width: thin;
+    scrollbar-color: hsl(var(--muted-foreground) / 0.3) transparent;
+  }
+
+  /* Webkit scrollbar styling for better appearance */
+  .node-content-area::-webkit-scrollbar {
+    width: 8px;
+  }
+
+  .node-content-area::-webkit-scrollbar-track {
+    background: transparent;
+  }
+
+  .node-content-area::-webkit-scrollbar-thumb {
+    background: hsl(var(--muted-foreground) / 0.3);
+    border-radius: 4px;
+  }
+
+  .node-content-area::-webkit-scrollbar-thumb:hover {
+    background: hsl(var(--muted-foreground) / 0.5);
+  }
+
+  /* Calendar icon styling */
+  :global(.calendar-icon) {
+    opacity: 0.8;
+  }
+
+  /* Left arrow rotation for previous button */
+  :global(.rotate-left) {
+    transform: rotate(180deg);
   }
 </style>
